feat: Implement pagination across various tables and enhance UI components
- Added pagination functionality to OrdersTable, PackagesTable, PurshasesTable, ProductsTable, and UsersTable components.
- Introduced a new Pagination component for consistent pagination UI.
- Updated the formatDayMonth function to display month in short format.
- Enhanced ProductsFilters to include ProductsColumnsSelector for better column visibility management.
- Improved styling and layout for better user experience in tables.

+++ apps/admin/src/lib/format-date.ts
@@ -66,7 +66,7 @@ export const formatDayMonth = (date: string | Date): string => {
   if (isNaN(dateObj.getTime())) return '';
   return dateObj.toLocaleDateString('es-ES', {
     day: '2-digit',
-    month: 'long'
+    month: 'short'
   });
 };
